Convert useDropiesAround to context provider
