+++ src/components/map/DropyMap.js
@@ -18,6 +18,7 @@ import MapLoadingOverlay from '../overlays/MapLoadingOverlay';
 import DebugText from '../other/DebugText';
 import FadeInWrapper from '../effect/FadeInWrapper';
 import EnergyPopup from '../overlays/EnergyPopup';
+import useDropiesAroundSocket from '../../hooks/useDropiesAroundSocket';
 import RetrievedDropyMapMarker from './RetrievedDropyMapMarker';
 import Sonar from './Sonar';
 import DropyMapMarker from './DropyMapMarker';
@@ -25,8 +26,6 @@ import MapDebugger from './MapDebugger';
 import OSMapView from './OSMapView';
 
 const DropyMap = ({
-  dropiesAround,
-  retrieveDropy,
   museumVisible,
   selectedDropyIndex = null,
   retrievedDropies = null,
@@ -34,11 +33,18 @@ const DropyMap = ({
   const navigation = useNavigation();
 
   const { sendBottomAlert, sendAlert } = useOverlay();
+
+  const {
+    dropiesAround,
+    retrieveDropy,
+  } = useDropiesAroundSocket();
+
   const {
     userCoordinates,
     compassHeading,
     initialized: geolocationInitialized,
   } = useInitializedGeolocation();
+
   const { developerMode, setUser } = useCurrentUser();
 
   const [currentZoom, setCurrentZoom] = useState(0);
